+++ stories/layout/Expandable.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
 import { action } from 'storybook/actions'
-import { ExpandableUncontrolled } from '../../src/components/layout-and-navigation/Expandable'
+import { ExpandableUncontrolled } from '../../src/components/layout/Expandable'
 import { range } from '../../src/utils/array'
 
 const meta = {